First Scala.js Runner implementation
