+++ project/WeaverPlugin.scala
@@ -103,7 +103,7 @@ object WeaverPlugin extends AutoPlugin {
     "-Ywarn-unused:patvars",         // Warn if a variable bound in a pattern is unused.
     "-Ywarn-unused:privates",        // Warn if a private member is unused.
     "-Ywarn-value-discard",          // Warn when non-Unit expression results are unused.
-    "-Xfatal-warnings"               // Fail the compilation if there are any warnings.
+    // "-Xfatal-warnings"               // Fail the compilation if there are any warnings.
   )
 
   lazy val compilerOptions2_12_Only =
